feat(trackers): bascule complete sur le mecanisme engine + suppression du hardcode par-tracker
- 20 JSON migres sur engine (unit3d/gazelle) avec surcharges locales

- server.ts: suppression de knownTrackerFields/knownUnit3dFields, des patches login et de COOKIE_ONLY_TRACKERS

- db.ts: applyEnginePreset applique a la lecture (loadTrackerConfigsFromDb)

- trackerTemplates.ts: presets ajustes (regle majorite), torrentleech/mam retires

- redacted: otpField qrcode_confirm conserve en JSON (varie par site, hors preset)
