+++ config/trackers/bitporn.json
@@ -3,38 +3,14 @@
   "name": "BitPorn",
   "baseUrl": "https://bitporn.eu",
   "enabled": false,
+  "engine": "unit3d",
   "login": {
     "url": "login",
     "method": "POST",
     "contentType": "form",
-    "preStep": {
-      "url": "login",
-      "extract": {
-        "_csrf": {
-          "regex": "name=\"_token\"\\s+value=\"(?<value>[^\"]+)\""
-        }
-      },
-      "includeHiddenInputs": true
-    },
-    "body": {
-      "_token": "{{_csrf}}",
-      "username": "{{username}}",
-      "password": "{{password}}",
-      "remember": "on"
-    },
-    "failurePatterns": [
-      "auth-form__form",
-      "type=\"password\"",
-      "name=\"password\"",
-      "These credentials do not match",
-      "Bejelentkezés"
-    ],
     "cookieOnly": true
   },
   "fetch": {
-    "url": "/",
-    "mode": "browser",
-    "responseType": "html",
     "fields": {
       "uploadedBytes": {
         "regex": "ratio-bar__uploaded[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)",
@@ -44,25 +20,10 @@
         "regex": "ratio-bar__downloaded[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)",
         "transform": "bytes"
       },
-      "ratio": {
-        "regex": "ratio-bar__ratio[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,]+)",
-        "transform": "number"
-      },
-      "seeding": {
-        "regex": "ratio-bar__seeding[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>\\d+)",
-        "transform": "integer"
-      },
-      "seedBonus": {
-        "regex": "ratio-bar__points[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,\\u202f]+)",
-        "transform": "string"
-      },
       "bufferBytes": {
         "regex": "ratio-bar__buffer[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)",
         "transform": "bytes"
       }
     }
-  },
-  "dashboard": {
-    "byteUnit": "binary"
   }
 }
