+++ config/trackers/phoenixproject.json
@@ -3,7 +3,7 @@
   "name": "Phoenix Project",
   "baseUrl": "https://phoenixproject.app",
   "enabled": true,
-  "curlBinary": "curl_firefox135",
+  "engine": "gazelle",
   "login": {
     "url": "login.php",
     "method": "POST",
@@ -21,9 +21,6 @@
     ]
   },
   "fetch": {
-    "url": "index.php",
-    "mode": "browser",
-    "responseType": "html",
     "fields": {
       "uploadedBytes": {
         "regex": "id=\"stats_seeding\"[\\s\\S]{0,200}?title=\"(?<value>[^\"]+)\"",
@@ -47,5 +44,5 @@
       }
     }
   },
-  "dashboard": { "byteUnit": "binary" }
+  "curlBinary": "curl_firefox135"
 }
