+++ config/trackers/redacted.json
@@ -3,44 +3,15 @@
   "name": "Redacted",
   "baseUrl": "https://redacted.sh",
   "enabled": true,
+  "engine": "gazelle",
   "login": {
     "url": "login.php",
     "method": "POST",
     "contentType": "form",
-    "preStep": {
-      "url": "login.php",
-      "extract": {},
-      "includeHiddenInputs": true
-    },
-    "body": {
-      "username": "{{username}}",
-      "password": "{{password}}",
-      "keeplogged": "1",
-      "login": "Log in"
-    },
-    "otpField": "qrcode_confirm",
-    "failurePatterns": [
-      "type=\"password\"",
-      "name=\"password\""
-    ]
+    "otpField": "qrcode_confirm"
   },
   "fetch": {
-    "url": "index.php",
-    "mode": "browser",
-    "responseType": "html",
     "fields": {
-      "downloadedBytes": {
-        "regex": "id=\"stats_leeching\"[^>]*title=\"Downloaded: (?<value>[^\"]+)\"",
-        "transform": "bytes"
-      },
-      "uploadedBytes": {
-        "regex": "id=\"stats_seeding\"[^>]*title=\"Uploaded: (?<value>[^\"]+)\"",
-        "transform": "bytes"
-      },
-      "ratio": {
-        "regex": "id=\"stats_ratio\"[^>]*title=\"Ratio: (?<value>[^\"]+)\"",
-        "transform": "number"
-      },
       "requiredRatio": {
         "regex": "id=\"stats_required\"[^>]*title=\"Required Ratio: (?<value>[^\"]+)\"",
         "transform": "number"
@@ -50,8 +21,5 @@
         "transform": "integer"
       }
     }
-  },
-  "dashboard": {
-    "byteUnit": "binary"
   }
 }
