+++ config/trackers/tr4ker.json
@@ -16,7 +16,8 @@
       "type=\"password\"",
       "name=\"password\"",
       "href=\"/login\"",
-      "aria-label=\"Connexion\""
+      "aria-label=\"Connexion\"",
+      "Inscription"
     ],
     "cookieOnly": true
   },
@@ -25,10 +26,21 @@
     "mode": "browser",
     "responseType": "html",
     "fields": {
-      "ratio":           { "regex": "RATIO\\s*:?\\s*(?<value>[\\d\\s.,]+)",                                             "transform": "number" },
-      "uploadedBytes":   { "regex": ">UPLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<",   "transform": "bytes" },
-      "downloadedBytes": { "regex": ">DOWNLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<", "transform": "bytes" }
+      "ratio": {
+        "regex": "RATIO\\s*:?\\s*(?<value>[\\d\\s.,]+)",
+        "transform": "number"
+      },
+      "uploadedBytes": {
+        "regex": ">UPLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<",
+        "transform": "bytes"
+      },
+      "downloadedBytes": {
+        "regex": ">DOWNLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<",
+        "transform": "bytes"
+      }
     }
   },
-  "dashboard": { "byteUnit": "decimal" }
+  "dashboard": {
+    "byteUnit": "decimal"
+  }
 }
