+++ config/trackers/yggreborn.json
@@ -15,8 +15,8 @@
       "type=\"password\"",
       "name=\"password\"",
       "name=\"identifier\"",
-      "Connexion à ton compte",
-      "cf-turnstile"
+      "cf-turnstile",
+      "Connexion à ton compte"
     ],
     "cookieOnly": true
   },
@@ -25,9 +25,17 @@
     "mode": "browser",
     "responseType": "html",
     "fields": {
-      "uploadedBytes":   { "regex": "(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Upload<",   "transform": "bytes" },
-      "downloadedBytes": { "regex": "(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Download<", "transform": "bytes" }
+      "uploadedBytes": {
+        "regex": "(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Upload<",
+        "transform": "bytes"
+      },
+      "downloadedBytes": {
+        "regex": "(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Download<",
+        "transform": "bytes"
+      }
     }
   },
-  "dashboard": { "byteUnit": "decimal" }
+  "dashboard": {
+    "byteUnit": "decimal"
+  }
 }
