+++ src/server.ts
@@ -41,6 +41,7 @@ import {
   listTrackerSchedules,
   loadCredentialsFromDb,
   loadTrackerConfigsFromDb,
+  loadRawTrackerConfigsFromDb,
   loadTrackerDefinitionFile,
   loadDefaultTrackerDefinition,
   saveStatSnapshots,
@@ -273,484 +274,13 @@ let betaQbitLastRefresh: string | null = null;
 // rafraîchissement automatique par intervalle (en mémoire, comme betaQbitStats).
 const qbitClientLastScan = new Map<string, number>();
 
-const unit3dFields = knownUnit3dFields();
-const gazelleStatsFields = knownGazelleStatsFields();
 
-function knownUnit3dFields(): Record<string, FieldExtractor> {
-  return {
-    uploadedBytes: {
-      regex: 'ratio-bar__uploaded[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,]+\\s*[KMGTPE]?i?B)',
-      transform: 'bytes',
-    },
-    downloadedBytes: {
-      regex: 'ratio-bar__downloaded[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,]+\\s*[KMGTPE]?i?B)',
-      transform: 'bytes',
-    },
-    ratio: {
-      regex: 'ratio-bar__ratio[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,]+)',
-      transform: 'number',
-    },
-    seeding: {
-      regex: 'ratio-bar__seeding[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>\\d+)',
-      transform: 'integer',
-    },
-    leeching: {
-      regex: 'ratio-bar__leeching[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>\\d+)',
-      transform: 'integer',
-    },
-    seedBonus: {
-      regex: 'ratio-bar__points[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,\\u202f]+)',
-      transform: 'string',
-    },
-    bufferBytes: {
-      regex: 'ratio-bar__buffer[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>[\\d\\s.,]+\\s*[KMGTPE]?i?B)',
-      transform: 'bytes',
-    },
-    tokens: {
-      regex: 'ratio-bar__tokens[\\s\\S]*?<i[^>]*>[\\s\\S]*?</i>\\s*(?<value>\\d+)',
-      transform: 'integer',
-    },
-    // MP non lus : UNIT3D affiche une sphère animée (<animate>) près de l'icône envelope
-    // dans le bloc "conversations" quand il y a des MP non lus, rien sinon.
-    // Pas de compteur disponible → on capture la durée de l'animation ("1s") comme
-    // signal de présence ; unreadMessagesCount() mappe toute string non vide → 1.
-    unreadMessages: {
-      regex: 'Boîte de réception(?:(?!<\\/a>)[\\s\\S])*?<animate[^>]*?dur="(?<value>[^"]+)"',
-      transform: 'string',
-    },
-  };
-}
 
-function knownGazelleStatsFields(): Record<string, FieldExtractor> {
-  return {
-    seedBonus: {
-      regex: '>Credits</a>:\\s*</td>[\\s\\S]{0,140}?<span[^>]*class=[\'"]stat[\'"][^>]*>\\s*(?<value>[\\d\\s.,]+)',
-      transform: 'string',
-    },
-    uploadedBytes: {
-      regex: '>Up</a>:\\s*</td>[\\s\\S]{0,160}?<span[^>]*class=[\'"]stat[\'"][^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-      transform: 'bytes',
-    },
-    downloadedBytes: {
-      regex: '>Down</a>:\\s*</td>[\\s\\S]{0,160}?<span[^>]*class=[\'"]stat[\'"][^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-      transform: 'bytes',
-    },
-    ratio: {
-      regex: '>Ratio</a>:\\s*</td>[\\s\\S]{0,220}?<span[^>]*class=[\'"]r\\d+[\'"][^>]*>\\s*(?<value>\\d[\\d\\s.,]*)',
-      transform: 'number',
-    },
-  };
-}
-
-const knownTrackerFields: Record<string, {
-  fetchUrl?: string;
-  mode?: 'http' | 'browser';
-  byteUnit?: 'binary' | 'decimal';
-  ratioless?: boolean;
-  fields: Record<string, FieldExtractor>;
-}> = {
-  hdonly: {
-    fetchUrl: 'index.php',
-    byteUnit: 'decimal',
-    ratioless: true,
-    fields: {
-      uploadedBytes: {
-        regex: 'Envoy[\\s\\S]{0,160}?(?<value>[\\d\\s.,]+\\s*[KMGTPE]?i?B)',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'Re[\\s\\S]{0,160}?(?<value>[\\d\\s.,]+\\s*[KMGTPE]?i?B)',
-        transform: 'bytes',
-      },
-      // MP non lus : skin Gazelle identique à HDForever (gazelle-notification +
-      // data-notification-type="Inbox"). A confirmer par Aerya si skin différente.
-      unreadMessages: {
-        regex: 'data-notification-type=[\'"]Inbox[\'"][^>]*>[^<]*?(?<value>\\d+|\\bun) nouveau',
-        transform: 'string',
-      },
-    },
-  },
-  hdforever: {
-    fetchUrl: 'index.php',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: 'class="stat tooltip up" title="(?<value>[^"]+)"',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'class="stat tooltip dl" title="(?<value>[^"]+)"',
-        transform: 'bytes',
-      },
-      ratio: {
-        regex: 'stats_ratio[^>]*>Ratio[^<]*<[^>]*><span class="tooltip r\\d+" title="(?<value>[^"]+)"',
-        transform: 'number',
-      },
-      seedBonus: {
-        regex: 'action=rate[^>]+>(?<value>[\\d,]+)<',
-        transform: 'string',
-      },
-      // MP non lus : skin Gazelle "gazelle-notification" + data-notification-type="Inbox".
-      // FR : "Vous avez 2 nouveaux messages" mais "Vous avez un nouveau message" (singulier
-      // = mot, pas chiffre). On capte \d+ OU "un" UNIQUEMENT s'il est suivi de " nouveau"
-      // et precede d'une frontiere de mot (\b) -> evite "aucun" ou un "un" egare.
-      // transform 'string' pour preserver "un" ; le front mappe un/une -> 1.
-      unreadMessages: {
-        regex: 'data-notification-type=[\'"]Inbox[\'"][^>]*>[^<]*?(?<value>\\d+|\\bun) nouveau',
-        transform: 'string',
-      },
-    },
-  },
-  theoldschool: {
-    fetchUrl: '/',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  generationfree: {
-    fetchUrl: '/',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  teamflix: {
-    fetchUrl: '/',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  g3mini: {
-    fetchUrl: '/',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  seedpool: {
-    fetchUrl: '/',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  sextorrent: {
-    fetchUrl: '/',
-    byteUnit: 'binary',
-    fields: unit3dFields,
-  },
-  bitporn: {
-    fetchUrl: '/',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: {
-      // BitPorn enrobe la valeur dans un <a href=".../uploads"> et utilise un espace
-      // insecable ( ) -> on ancre sur la classe et on capture juste nombre+unite.
-      uploadedBytes: {
-        regex: 'ratio-bar__uploaded[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'ratio-bar__downloaded[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)',
-        transform: 'bytes',
-      },
-      ratio: unit3dFields.ratio,
-      seeding: unit3dFields.seeding,
-      seedBonus: unit3dFields.seedBonus,
-      bufferBytes: {
-        regex: 'ratio-bar__buffer[\\s\\S]{0,260}?(?<value>\\d[\\d.,]*(?:&nbsp;|&#160;|&#xa0;|[\\s\\u00a0])*[KMGTPE]i?B)',
-        transform: 'bytes',
-      },
-    },
-  },
-  exoticaz: {
-    fetchUrl: '/',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes:   { regex: 'fa-arrow-up[\\s\\S]{0,40}?</i>\\s*(?<value>[\\d.,]+\\s*[KMGTPE]?i?B)',   transform: 'bytes' },
-      downloadedBytes: { regex: 'fa-arrow-down[\\s\\S]{0,40}?</i>\\s*(?<value>[\\d.,]+\\s*[KMGTPE]?i?B)', transform: 'bytes' },
-      ratio:           { regex: 'fa-signal[\\s\\S]{0,40}?</i>\\s*(?<value>[\\d.,]+)',                     transform: 'number' },
-      bufferBytes:     { regex: 'fa-database[\\s\\S]{0,40}?</i>\\s*(?<value>[\\d.,]+\\s*[KMGTPE]?i?B)',    transform: 'bytes' },
-      seeding:         { regex: 'Seeding:</a>\\s*(?<value>\\d+)',                                          transform: 'integer' },
-      seedBonus:       { regex: 'Bonus:</a>\\s*(?<value>[\\d.,]+)',                                        transform: 'string' },
-    },
-  },
-  abnormal: {
-    fetchUrl: '/',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: 'Up\\s*:[\\s\\S]{0,360}?text-green[\\s\\S]{0,120}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B))',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'Down\\s*:[\\s\\S]{0,360}?text-green[\\s\\S]{0,120}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B))',
-        transform: 'bytes',
-      },
-      ratio: {
-        regex: 'Ratio\\s*:[\\s\\S]{0,360}?text-green[\\s\\S]{0,120}?(?<value>[\\d\\s.,]+)',
-        transform: 'number',
-      },
-      seedBonus: {
-        regex: "Choco's\\s*:[\\s\\S]{0,360}?text-green[\\s\\S]{0,120}?(?<value>[\\d\\s.,]+)",
-        transform: 'string',
-      },
-      // MP : notification maison "Vous avez N nouveau(x) message(s)." (signal ephemere,
-      // efface une fois "Voir"/"Je m'en fous" clique cote ABN). Capte \d+ OU "un" (au cas
-      // ou ABN ecrit le singulier en lettres), ancre sur "nouveau(x) message", \b anti-"aucun".
-      unreadMessages: {
-        regex: '(?<value>\\d+|\\bun) nouveau\\(x\\) message',
-        transform: 'string',
-      },
-    },
-  },
-  yggreborn: {
-    fetchUrl: 'account/',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: '(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Upload<',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: '(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</div>[\\s\\S]{0,180}?>Download<',
-        transform: 'bytes',
-      },
-    },
-  },
-  tr4ker: {
-    fetchUrl: '/',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    fields: {
-      ratio: {
-        regex: 'RATIO\\s*:?\\s*(?<value>[\\d\\s.,]+)',
-        transform: 'number',
-      },
-      uploadedBytes: {
-        regex: '>UPLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: '>DOWNLOAD<[\\s\\S]{0,180}?_statValue[^>]*>\\s*(?<value>\\d[\\d\\s.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<',
-        transform: 'bytes',
-      },
-    },
-  },
-  lacale: {
-    fetchUrl: 'profile',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: '\\\\?"uploaded\\\\?"\\s*:\\s*(?<value>\\d+)',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: '\\\\?"downloaded\\\\?"\\s*:\\s*(?<value>\\d+)',
-        transform: 'bytes',
-      },
-      seedBonus: {
-        regex: '\\\\?"bonusPoints\\\\?"\\s*:\\s*(?<value>\\d+)',
-        transform: 'string',
-      },
-    },
-  },
-  crazyspirits: {
-    fetchUrl: '/',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: {
-      downloadedBytes: {
-        regex: '/dl\\.png"[\\s\\S]{0,160}?<font[^>]*>\\s*(?<value>[\\d.,]+\\s*[KMGTPE]?i?B)',
-        transform: 'bytes',
-      },
-      uploadedBytes: {
-        regex: '/up\\.png"[\\s\\S]{0,160}?<font[^>]*>\\s*(?<value>[\\d.,]+\\s*[KMGTPE]?i?B)',
-        transform: 'bytes',
-      },
-      seedBonus: {
-        regex: 'Crazy Bonus\\s*<a[^>]*>\\s*(?<value>[\\d\\s.,]+)',
-        transform: 'string',
-      },
-    },
-  },
-  tigersdl: {
-    fetchUrl: 'mybonus.php',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: 'title=[\'"]Partager[\'"][\\s\\S]{0,180}?<font[^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</font>',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'title=[\'"][^\'"]*charg[^\'"]*[\'"][\\s\\S]{0,180}?<font[^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</font>',
-        transform: 'bytes',
-      },
-      seedBonus: {
-        regex: 'Votre solde[\\s\\S]{0,180}?score-points[^>]*>\\s*(?<value>[\\d\\s.,]+)',
-        transform: 'string',
-      },
-      seeding: {
-        regex: '(?:Nombres de Torrents que vous avez en seed\\s*:\\s*|title=[\'"]Seeding[\'"][\\s\\S]{0,120}?<b>\\s*)(?<value>\\d+)',
-        transform: 'integer',
-      },
-    },
-  },
-  mam: {
-    fetchUrl: 'u/',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: {
-      uploadedBytes: {
-        regex: 'Uploaded[\\s\\S]{0,160}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'Downloaded[\\s\\S]{0,160}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-        transform: 'bytes',
-      },
-      ratio: {
-        regex: 'Share ratio[\\s\\S]{0,160}?(?<value>\\d[\\d\\s.,]*)',
-        transform: 'number',
-      },
-      seedBonus: {
-        regex: 'Bonus[\\s\\S]{0,160}?(?<value>[\\d\\s.,]+)',
-        transform: 'string',
-      },
-    },
-  },
-  milkie: {
-    fetchUrl: 'browse',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: {
-      uploadedBytes: {
-        regex: 'keyboard_arrow_up[\\s\\S]{0,140}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'keyboard_arrow_down[\\s\\S]{0,140}?(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))',
-        transform: 'bytes',
-      },
-    },
-  },
-  speedapp: {
-    fetchUrl: 'profile',
-    byteUnit: 'decimal',
-    fields: {
-      uploadedBytes: {
-        regex: 'Uploaded[\\s\\S]{0,120}?<dd[^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</dd>',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: 'Downloaded[\\s\\S]{0,120}?<dd[^>]*>\\s*(?<value>[\\d\\s.,]+\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*</dd>',
-        transform: 'bytes',
-      },
-      seedTimeDays: {
-        regex: 'Seed time[\\s\\S]{0,120}?<dd[^>]*>\\s*(?<value>\\d[\\d\\s.,]*)\\s*days',
-        transform: 'number',
-      },
-    },
-  },
-  kufirc: {
-    fetchUrl: 'index.php',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: gazelleStatsFields,
-  },
-  happyfappy: {
-    fetchUrl: 'index.php',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: gazelleStatsFields,
-  },
-  empornium: {
-    fetchUrl: 'index.php',
-    mode: 'browser',
-    byteUnit: 'binary',
-    fields: gazelleStatsFields,
-  },
-  nostradamus: {
-    fetchUrl: 'activity',
-    mode: 'browser',
-    byteUnit: 'decimal',
-    ratioless: true,
-    fields: {
-      uploadedBytes: {
-        regex: '>\\s*Upload total\\s*<[\\s\\S]{0,400}?>\\s*(?<value>\\d[\\d.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<',
-        transform: 'bytes',
-      },
-      downloadedBytes: {
-        regex: '>\\s*Download total\\s*<[\\s\\S]{0,400}?>\\s*(?<value>\\d[\\d.,]*\\s*(?:[KMGTPE](?:i?B|io|o)|B|o))\\s*<',
-        transform: 'bytes',
-      },
-      points: {
-        regex: 'hero-banknotes[\\s\\S]*?sidebar-account-stat__value[^>]*>\\s*(?<value>[\\d\\s.,]+)',
-        transform: 'integer',
-      },
-      rate: {
-        regex: 'hero-bolt[\\s\\S]*?sidebar-account-stat__value[^>]*>\\s*(?<value>[\\d\\s.,]+)',
-        transform: 'number',
-      },
-    },
-  },
-};
-
-// Trackers dont le login ne peut pas etre automatise (CAPTCHA, Cloudflare Turnstile,
-// plafond de sessions...) : on force le mode "cookie uniquement".
-const COOKIE_ONLY_TRACKERS = new Set([
-  'mam',          // plafond de sessions + login bloque
-  'seedpool',     // CAPTCHA au login (UNIT3D)
-  'bitporn',      // Cloudflare
-  'exoticaz',     // CAPTCHA
-  'crazyspirits', // Cloudflare Turnstile
-  'tr4ker',       // Cloudflare + login SPA
-  'yggreborn',    // Cloudflare Turnstile
-]);
-
-/**
- * Applique le preset moteur (ENGINE_TEMPLATES) à une config qui déclare `engine`.
- * RÈGLE D'OR : le JSON du tracker gagne TOUJOURS sur le preset, champ par champ.
- * - login / dashboard / curlBinary / ratioless : le bloc du JSON, s'il est présent,
- *   remplace celui du preset (au niveau du bloc login, on fusionne clé par clé pour
- *   qu'un tracker puisse surcharger seulement `url` sans réécrire body/preStep/…).
- * - fetch.fields : MERGE field par field (preset d'abord, JSON ensuite) pour qu'un
- *   site au skin modifié puisse surcharger 2 fields sur 9 sans réécrire les 7 autres.
- * - fetch (hors fields) : le JSON gagne champ par champ.
- * Un tracker sans `engine`, ou avec un `engine` inconnu, est renvoyé inchangé.
- * Idempotent : ré-appliquer ne change rien.
- */
-function applyEnginePreset(config: TrackerConfig): TrackerConfig {
-  if (!config.engine) return config;
-  const tpl = getEngineTemplate(config.engine);
-  if (!tpl) return config;
-  const preset = tpl.preset;
-
-  // login : preset comme base, le JSON surcharge clé par clé.
-  const mergedLogin = { ...preset.login, ...config.login };
-
-  // fetch : preset comme base, le JSON surcharge clé par clé, SAUF fields qui mergent.
-  const mergedFields = {
-    ...(preset.fetch?.fields ?? {}),
-    ...(config.fetch?.fields ?? {}),
-  };
-  const mergedFetch = { ...preset.fetch, ...config.fetch, fields: mergedFields };
-
-  // dashboard : preset comme base, JSON surcharge clé par clé.
-  const mergedDashboard =
-    config.dashboard || preset.dashboard
-      ? { ...(preset.dashboard ?? {}), ...(config.dashboard ?? {}) }
-      : undefined;
-
-  return {
-    ...config,
-    curlBinary: config.curlBinary ?? preset.curlBinary,
-    ratioless: config.ratioless ?? preset.ratioless,
-    login: mergedLogin,
-    fetch: mergedFetch,
-    ...(mergedDashboard ? { dashboard: mergedDashboard } : {}),
-  };
-}
 
 function normalizeTrackerConfigs(): TrackerConfig[] {
-  const trackers = loadTrackerConfigsFromDb();
+  // Synchro base/image sur les configs BRUTES (sans preset moteur). Le preset moteur
+  // est applique ensuite, a la lecture finale (loadTrackerConfigsFromDb).
+  const trackers = loadRawTrackerConfigsFromDb();
   for (const tracker of trackers) {
     let changed = false;
     // Les champs techniques (login, fetch, curlBinary) font autorite cote image :
@@ -775,296 +305,20 @@ function normalizeTrackerConfigs(): TrackerConfig[] {
         tracker.curlBinary = definition.curlBinary;
         changed = true;
       }
-    }
-    // Mécanisme `engine` : si le tracker déclare une famille de moteur, on applique
-    // le preset (ENGINE_TEMPLATES) — le JSON gagne toujours sur le preset (cf.
-    // applyEnginePreset). Tant qu'aucun JSON ne porte `engine`, ceci est un no-op.
-    // À terme, ce mécanisme remplace les patches par-tracker ci-dessous.
-    if (tracker.engine) {
-      const merged = applyEnginePreset(tracker);
-      if (JSON.stringify(merged.login) !== JSON.stringify(tracker.login)) {
-        tracker.login = merged.login;
+      // `engine` fait aussi autorite cote image : sans cette synchro, une base
+      // existante ne recevrait jamais la bascule vers le mecanisme engine (le tracker
+      // resterait sans engine, donc sans preset applique, fetch/login incomplets).
+      if (tracker.engine !== definition.engine) {
+        tracker.engine = definition.engine;
         changed = true;
       }
-      if (JSON.stringify(merged.fetch) !== JSON.stringify(tracker.fetch)) {
-        tracker.fetch = merged.fetch;
+      // dashboard / ratioless : surcharges declarees dans le JSON image, a refleter.
+      if (JSON.stringify(tracker.dashboard) !== JSON.stringify(definition.dashboard)) {
+        tracker.dashboard = definition.dashboard;
         changed = true;
       }
-      if (JSON.stringify(merged.dashboard) !== JSON.stringify(tracker.dashboard)) {
-        tracker.dashboard = merged.dashboard;
-        changed = true;
-      }
-      if (merged.curlBinary !== tracker.curlBinary) {
-        tracker.curlBinary = merged.curlBinary;
-        changed = true;
-      }
-      if (merged.ratioless !== tracker.ratioless) {
-        tracker.ratioless = merged.ratioless;
-        changed = true;
-      }
-    }
-    const isHdOnlyLikeTracker = ['hdonly', 'hdforever'].includes(tracker.id);
-    const isUnit3dTracker = ['theoldschool', 'generationfree', 'teamflix', 'g3mini', 'seedpool', 'bitporn', 'sextorrent'].includes(tracker.id);
-    if (tracker.id === 'hdonly' && tracker.login.failurePatterns.includes('login.php')) {
-      tracker.login.failurePatterns = tracker.login.failurePatterns
-        .filter(pattern => pattern !== 'login.php');
-      changed = true;
-    }
-    if (isHdOnlyLikeTracker) {
-      tracker.login.preStep = {
-        url: 'login.php',
-        extract: {},
-        includeHiddenInputs: true,
-      };
-      tracker.login.failurePatterns = [
-        ...new Set([
-          ...tracker.login.failurePatterns,
-          'type="password"',
-          'href="login.php"',
-          'Entrer</a>',
-        ]),
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'hdonly') {
-      tracker.login.body = {
-        username: '{{username}}',
-        password: '{{password}}',
-        keeplogged: '1',
-        login: 'Se connecter',
-      };
-      changed = true;
-    }
-    if (isUnit3dTracker) {
-      tracker.login.preStep = {
-        url: tracker.login.preStep?.url ?? 'login',
-        includeHiddenInputs: true,
-        // CSRF robuste : accepte l'input cache <input name="_token" value="..."> OU
-        // le <meta name="csrf-token" content="..."> du <head> (selon le rendu UNIT3D).
-        extract: {
-          _csrf: { regex: '(?:name="_token"[^>]*?\\svalue="|name="csrf-token"[^>]*?\\scontent=")(?<value>[^"]+)"' },
-        },
-      };
-      tracker.login.body = {
-        _token: '{{_csrf}}',
-        username: '{{username}}',
-        password: '{{password}}',
-        remember: 'on',
-      };
-      // Jeu CURÉ et fiable : ces marqueurs n'apparaissent QUE sur la page de login
-      // UNIT3D, jamais sur une page connectée. On ECRASE (pas de merge) pour eliminer
-      // les patterns trop larges venant des JSON (ex: "Bejelentkezés"/"Se connecter"
-      // qui figurent dans un menu meme connecte -> faux "non authentifie").
-      tracker.login.failurePatterns = [
-        'auth-form__form',
-        'type="password"',
-        'name="password"',
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'abnormal') {
-      if (tracker.login.url !== 'Home/Login') {
-        tracker.login.url = 'Home/Login';
-        changed = true;
-      }
-      tracker.login.preStep = {
-        ...(tracker.login.preStep ?? { url: 'Home/Login', extract: {} }),
-        url: 'Home/Login',
-        includeHiddenInputs: true,
-      };
-      tracker.login.failurePatterns = [
-        ...new Set([
-          ...tracker.login.failurePatterns,
-          'id="account"',
-          'type="password"',
-          'Connexion - ABN',
-        ]),
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'nostradamus') {
-      if (tracker.login.url !== 'sign-in') {
-        tracker.login.url = 'sign-in';
-      }
-      tracker.login.body = {
-        password: '{{password}}',
-      };
-      tracker.login.failurePatterns = [
-        'type="password"',
-        'private-key-input',
-        'name="password"',
-        'name="username"',
-        'Se connecter',
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'tr4ker') {
-      tracker.login.failurePatterns = [
-        ...new Set([
-          ...tracker.login.failurePatterns,
-          'href="/login"',
-          'aria-label="Connexion"',
-          'Inscription',
-        ]),
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'phoenixproject') {
-      if (tracker.curlBinary !== 'curl_firefox135') {
-        tracker.curlBinary = 'curl_firefox135';
-        changed = true;
-      }
-      if (tracker.login.cookieOnly) {
-        tracker.login.cookieOnly = false;
-        changed = true;
-      }
-    }
-    if (tracker.id === 'tigersdl') {
-      if (tracker.login.url !== 'account-login.php') {
-        tracker.login.url = 'account-login.php';
-      }
-      tracker.login.body = {
-        username: '{{username}}',
-        password: '{{password}}',
-      };
-      tracker.login.failurePatterns = [
-        'account-login.php',
-        'type="password"',
-        'type=\'password\'',
-        'name="password"',
-        'name=\'password\'',
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'mam') {
-      if (tracker.login.url !== 'login.php?returnto=%2Fu%2F') {
-        tracker.login.url = 'login.php?returnto=%2Fu%2F';
-      }
-      tracker.login.preStep = {
-        url: 'login.php?returnto=%2Fu%2F',
-        extract: {},
-        includeHiddenInputs: true,
-      };
-      tracker.login.body = {
-        email: '{{username}}',
-        password: '{{password}}',
-        rememberMe: 'yes',
-        returnto: '/u/',
-      };
-      tracker.login.failurePatterns = [
-        'Not logged in!',
-        'Login | My Anonamouse',
-        'name="password"',
-        'name=\'password\'',
-        'loginIssueBlock',
-      ];
-      changed = true;
-    }
-    // Trackers a CAPTCHA / Cloudflare Turnstile / anti-bot : le login formulaire est
-    // impossible a automatiser -> cookie de session uniquement (jamais de soumission
-    // de formulaire). Force le flag meme si la config en base est ancienne.
-    if (COOKIE_ONLY_TRACKERS.has(tracker.id) && !tracker.login.cookieOnly) {
-      tracker.login.cookieOnly = true;
-      changed = true;
-    }
-    if (tracker.id === 'milkie') {
-      if (tracker.login.url !== 'auth/signin') {
-        tracker.login.url = 'auth/signin';
-      }
-      tracker.login.body = {
-        email: '{{username}}',
-        password: '{{password}}',
-      };
-      tracker.login.failurePatterns = [
-        '/auth/signin',
-        'type="password"',
-        'type=\'password\'',
-        'name="password"',
-        'name=\'password\'',
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'speedapp') {
-      if (tracker.login.url !== 'fr/connexion?locale=fr') {
-        tracker.login.url = 'fr/connexion?locale=fr';
-      }
-      tracker.login.preStep = {
-        url: 'fr/connexion?locale=fr',
-        extract: {},
-        includeHiddenInputs: true,
-      };
-      tracker.login.body = {
-        username: '{{username}}',
-        password: '{{password}}',
-        _remember_me: 'on',
-      };
-      tracker.login.failurePatterns = [
-        'Connexion | SpeedApp',
-        'name="username"',
-        'name="password"',
-        'name=\'_csrf_token\'',
-      ];
-      changed = true;
-    }
-    if (['kufirc', 'happyfappy', 'empornium'].includes(tracker.id)) {
-      if (tracker.login.url !== 'login') {
-        tracker.login.url = 'login';
-      }
-      tracker.login.preStep = {
-        url: 'login',
-        extract: {},
-        includeHiddenInputs: true,
-      };
-      tracker.login.body = {
-        username: '{{username}}',
-        password: '{{password}}',
-        keeploggedin: '1',
-      };
-      tracker.login.failurePatterns = [
-        'Login ::',
-        'name="password"',
-        'id="login_button"',
-        'href="/login"',
-      ];
-      changed = true;
-    }
-    if (tracker.id === 'yggreborn') {
-      if (tracker.login.url !== 'login?next=/account/') {
-        tracker.login.url = 'login?next=/account/';
-      }
-      tracker.login.body = {
-        identifier: '{{username}}',
-        password: '{{password}}',
-      };
-      tracker.login.failurePatterns = [
-        'type="password"',
-        'name="password"',
-        'name="identifier"',
-        'cf-turnstile',
-        'Connexion à ton compte',
-      ];
-      changed = true;
-    }
-    const known = knownTrackerFields[tracker.id];
-    if (known) {
-      if (known.mode && tracker.fetch.mode !== known.mode) {
-        tracker.fetch.mode = known.mode;
-        changed = true;
-      }
-      if (known.fetchUrl && tracker.fetch.url !== known.fetchUrl) {
-        tracker.fetch.url = known.fetchUrl;
-        changed = true;
-      }
-      if (JSON.stringify(tracker.fetch.fields) !== JSON.stringify(known.fields)) {
-        tracker.fetch.fields = known.fields;
-        changed = true;
-      }
-      if (known.byteUnit && tracker.dashboard?.byteUnit !== known.byteUnit) {
-        tracker.dashboard = { ...(tracker.dashboard ?? {}), byteUnit: known.byteUnit };
-        changed = true;
-      }
-      if (typeof known.ratioless === 'boolean' && tracker.ratioless !== known.ratioless) {
-        tracker.ratioless = known.ratioless;
+      if (Boolean(tracker.ratioless) !== Boolean(definition.ratioless)) {
+        tracker.ratioless = definition.ratioless;
         changed = true;
       }
     }
